+++ crates/ruma-state-res/src/event_format.rs
@@ -1,6 +1,7 @@
 use js_int::int;
 use ruma_common::{
-    room_version_rules::EventFormatRules, CanonicalJsonObject, CanonicalJsonValue, ID_MAX_BYTES,
+    room_version_rules::EventFormatRules, CanonicalJsonObject, CanonicalJsonValue, RoomId,
+    ID_MAX_BYTES,
 };
 use serde_json::to_string as to_json_string;
 
@@ -48,55 +49,58 @@ pub fn check_pdu_format(pdu: &CanonicalJsonObject, rules: &EventFormatRules) -> 
         return Err("PDU is larger than maximum of {MAX_PDU_BYTES} bytes".to_owned());
     }
 
-    // Check the presence, type and length of the `sender`, `room_id` and `type` fields.
-    for field in &["sender", "room_id", "type"] {
-        let value = extract_string_field(pdu, field)?
-            .ok_or_else(|| format!("missing `{field}` field in PDU"))?;
+    // Check the presence, type and length of the `type` field.
+    let event_type = extract_required_string_field(pdu, "type")?;
 
-        if value.len() > ID_MAX_BYTES {
-            return Err(format!(
-                "invalid `{field}` field in PDU: \
-                 string length is larger than maximum of {ID_MAX_BYTES} bytes"
-            ));
-        }
-    }
+    // Check the presence, type and length of the `sender` field.
+    extract_required_string_field(pdu, "sender")?;
+
+    // Check the presence, type and length of the `room_id` field.
+    let room_id = (event_type != "m.room.create" || rules.require_room_create_room_id)
+        .then(|| extract_required_string_field(pdu, "room_id"))
+        .transpose()?;
 
     // Check the presence, type and length of the `event_id` field.
-    let event_id = extract_string_field(pdu, "event_id")?;
-
-    if rules.require_event_id && event_id.is_none() {
-        return Err("missing `event_id` field in PDU".to_owned());
-    }
-
-    if event_id.is_some_and(|event_id| event_id.len() > ID_MAX_BYTES) {
-        return Err(format!(
-            "invalid `event_id` field in PDU: \
-             string length is larger than maximum of {ID_MAX_BYTES} bytes"
-        ));
+    if rules.require_event_id {
+        extract_required_string_field(pdu, "event_id")?;
     }
 
     // Check the type and length of the `state_key` field.
-    if extract_string_field(pdu, "state_key")?
-        .is_some_and(|state_key| state_key.len() > ID_MAX_BYTES)
-    {
-        return Err(format!(
-            "invalid `state_key` field in PDU: \
-             string length is larger than maximum of {ID_MAX_BYTES} bytes"
-        ));
-    }
+    extract_optional_string_field(pdu, "state_key")?;
 
-    // Check the presence, type and length of the `auth_events` and `prev_events` fields.
-    for (field, max_value) in
-        &[("auth_events", MAX_AUTH_EVENTS_LENGTH), ("prev_events", MAX_PREV_EVENTS_LENGTH)]
-    {
-        let value = extract_array_field(pdu, field)?
-            .ok_or_else(|| format!("missing `{field}` field in PDU"))?;
+    // Check the presence, type and length of the `prev_events` field.
+    extract_required_array_field(pdu, "prev_events", MAX_PREV_EVENTS_LENGTH)?;
 
-        if value.len() > *max_value {
-            return Err(format!(
-                "invalid `{field}` field in PDU: \
-                 array length is larger than maximum of {max_value}"
-            ));
+    // Check the presence, type and length of the `auth_events` field.
+    let auth_events = extract_required_array_field(pdu, "auth_events", MAX_AUTH_EVENTS_LENGTH)?;
+
+    if !rules.allow_room_create_in_auth_events {
+        // The only case where the room ID should be missing is for m.room.create which shouldn't
+        // have any auth_events.
+        if let Some(room_id) = room_id {
+            let room_create_event_reference_hash = <&RoomId>::try_from(room_id.as_str())
+                .map_err(|e| format!("invalid `room_id` field in PDU: {e}"))?
+                .strip_sigil();
+
+            for event_id in auth_events {
+                let CanonicalJsonValue::String(event_id) = event_id else {
+                    return Err(format!(
+                        "unexpected format of array item in `auth_events` field in PDU: \
+                         expected string, got {event_id:?}"
+                    ));
+                };
+
+                let reference_hash = event_id.strip_prefix('$').ok_or(
+                    "unexpected format of array item in `auth_events` field in PDU: \
+                     string not beginning with the `$` sigil",
+                )?;
+
+                if reference_hash == room_create_event_reference_hash {
+                    return Err("invalid `auth_events` field in PDU: \
+                                cannot contain the `m.room.create` event ID"
+                        .to_owned());
+                }
+            }
         }
     }
 
@@ -119,16 +123,26 @@ pub fn check_pdu_format(pdu: &CanonicalJsonObject, rules: &EventFormatRules) -> 
     Ok(())
 }
 
-/// Extract the string field with the given name from the given canonical JSON object.
+/// Extract the optional string field with the given name from the given canonical JSON object.
 ///
-/// Returns `Ok(Some(value))` if the field is present and a string, `Ok(None)` if the fields is
-/// missing and an error if the field is present and not a string.
-fn extract_string_field<'a>(
+/// Returns `Ok(Some(value))` if the field is present and a valid string, `Ok(None)` if the field
+/// is missing and `Err(_)` if the field is not a string or its length is bigger than
+/// [`ID_MAX_BYTES`].
+fn extract_optional_string_field<'a>(
     object: &'a CanonicalJsonObject,
     field: &'a str,
 ) -> Result<Option<&'a String>, String> {
     match object.get(field) {
-        Some(CanonicalJsonValue::String(value)) => Ok(Some(value)),
+        Some(CanonicalJsonValue::String(value)) => {
+            if value.len() > ID_MAX_BYTES {
+                Err(format!(
+                    "invalid `{field}` field in PDU: \
+                     string length is larger than maximum of {ID_MAX_BYTES} bytes"
+                ))
+            } else {
+                Ok(Some(value))
+            }
+        }
         Some(value) => Err(format!(
             "unexpected format of `{field}` field in PDU: \
              expected string, got {value:?}"
@@ -137,21 +151,43 @@ fn extract_string_field<'a>(
     }
 }
 
-/// Extract the array field with the given name from the given canonical JSON object.
+/// Extract the required string field with the given name from the given canonical JSON object.
 ///
-/// Returns `Ok(Some(value))` if the field is present and an array, `Ok(None)` if the fields is
-/// missing and an error if the field is present and not a array.
-fn extract_array_field<'a>(
+/// Returns `Ok(value)` if the field is present and a valid string and `Err(_)` if the field is
+/// missing, not a string or its length is bigger than [`ID_MAX_BYTES`].
+fn extract_required_string_field<'a>(
     object: &'a CanonicalJsonObject,
     field: &'a str,
-) -> Result<Option<&'a [CanonicalJsonValue]>, String> {
+) -> Result<&'a String, String> {
+    extract_optional_string_field(object, field)?
+        .ok_or_else(|| format!("missing `{field}` field in PDU"))
+}
+
+/// Extract the required array field with the given name from the given canonical JSON object.
+///
+/// Returns `Ok(value)` if the field is present and a valid array or `Err(_)` if the field is
+/// missing, not an array or its length is bigger than the given value.
+fn extract_required_array_field<'a>(
+    object: &'a CanonicalJsonObject,
+    field: &'a str,
+    max_len: usize,
+) -> Result<&'a [CanonicalJsonValue], String> {
     match object.get(field) {
-        Some(CanonicalJsonValue::Array(value)) => Ok(Some(value)),
+        Some(CanonicalJsonValue::Array(value)) => {
+            if value.len() > max_len {
+                Err(format!(
+                    "invalid `{field}` field in PDU: \
+                     array length is larger than maximum of {max_len}"
+                ))
+            } else {
+                Ok(value)
+            }
+        }
         Some(value) => Err(format!(
             "unexpected format of `{field}` field in PDU: \
              expected array, got {value:?}"
         )),
-        None => Ok(None),
+        None => Err(format!("missing `{field}` field in PDU")),
     }
 }
 
@@ -227,6 +263,66 @@ mod tests {
             ],
             "redacts": "$some/old+event",
             "room_id": "!UcYsUzyxTGDxLBEvLy:example.org",
+            "sender": "@alice:example.com",
+            "signatures": {
+                "example.com": {
+                    "ed25519:key_version": "these86bytesofbase64signaturecoveressentialfieldsincludinghashessocancheckredactedpdus",
+                },
+            },
+            "type": "m.room.message",
+            "unsigned": {
+                "age": 4612,
+            }
+        });
+        from_json_value(pdu).unwrap()
+    }
+
+    /// Construct an `m.room.create` PDU valid for the event format of v12.
+    fn room_create_v12() -> CanonicalJsonObject {
+        let pdu = json!({
+            "auth_events": [],
+            "content": {
+                "room_version": "12",
+            },
+            "depth": 1,
+            "hashes": {
+                "sha256": "thishashcoversallfieldsincasethisisredacted",
+            },
+            "origin_server_ts": 1_838_188_000,
+            "prev_events": [],
+            "sender": "@alice:example.com",
+            "signatures": {
+                "example.com": {
+                    "ed25519:key_version": "these86bytesofbase64signaturecoveressentialfieldsincludinghashessocancheckredactedpdus",
+                },
+            },
+            "type": "m.room.create",
+            "unsigned": {
+                "age": 4612,
+            }
+        });
+        from_json_value(pdu).unwrap()
+    }
+
+    /// Construct a PDU valid for the event format of v12.
+    fn pdu_v12() -> CanonicalJsonObject {
+        let pdu = json!({
+            "auth_events": [
+                "$base64encodedeventid",
+                "$adifferenteventid",
+            ],
+            "content": {
+                "key": "value",
+            },
+            "depth": 12,
+            "hashes": {
+                "sha256": "thishashcoversallfieldsincasethisisredacted",
+            },
+            "origin_server_ts": 1_838_188_000,
+            "prev_events": [
+                "$base64encodedeventid",
+            ],
+            "room_id": "!roomcreatereferencehash",
             "sender": "@alice:example.com",
             "signatures": {
                 "example.com": {
@@ -325,5 +421,29 @@ mod tests {
         let mut pdu = pdu_v3();
         pdu.insert("depth".to_owned(), true.into());
         check_pdu_format(&pdu, &EventFormatRules::V3).unwrap_err();
+    }
+
+    #[test]
+    fn check_pdu_format_valid_room_create_v12() {
+        let pdu = room_create_v12();
+        check_pdu_format(&pdu, &EventFormatRules::V12).unwrap();
+    }
+
+    #[test]
+    fn check_pdu_format_valid_v12() {
+        let pdu = pdu_v12();
+        check_pdu_format(&pdu, &EventFormatRules::V12).unwrap();
+    }
+
+    #[test]
+    fn check_pdu_format_v12_with_room_create() {
+        let mut pdu = pdu_v12();
+        pdu.get_mut("auth_events")
+            .unwrap()
+            .as_array_mut()
+            .unwrap()
+            .push("$roomcreatereferencehash".to_owned().into());
+
+        check_pdu_format(&pdu, &EventFormatRules::V12).unwrap_err();
     }
 }